Implemented vertex compression for model writer
