+++ include/model_writer.h
@@ -1,0 +1,11 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <structs/models/model_data_str.h>
+
+// ====================== INTERNALLY DEFINED FUNCTIONS =======================
+
+// ====================== EXTERNALLY DEFINED FUNCTIONS =======================
+
+int comp_vec2(vec2 a, vec2 b);
+int comp_vec3(vec3 a, vec3 b);
+int double_buffer(void **buffer, size_t *buff_size, size_t unit_size);
