+++ src/model_loader.c
@@ -378,248 +378,3 @@ MODEL *load_model(char *path) {
   return model;
 }
 
-int write_model_obj(MODEL_DATA *md, char *path) {
-  // TODO Implement material saving
-  FILE *file = fopen(path, "w");
-  if (!file) {
-    return -1;
-  }
-
-  for (size_t i = 1; i < md->num_bones; i++) {
-    int parent = md->bones[i].parent;
-    if (parent != -1) {
-      parent--;
-    }
-    // Bones
-    fprintf(file, "b %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %d %d\n",
-                     md->bones[i].head[X],
-                     md->bones[i].head[Y],
-                     md->bones[i].head[Z],
-                     md->bones[i].base[X],
-                     md->bones[i].base[Y],
-                     md->bones[i].base[Z],
-                     md->bones[i].coordinate_matrix[0][0],
-                     md->bones[i].coordinate_matrix[0][1],
-                     md->bones[i].coordinate_matrix[0][2],
-                     md->bones[i].coordinate_matrix[1][0],
-                     md->bones[i].coordinate_matrix[1][1],
-                     md->bones[i].coordinate_matrix[1][2],
-                     md->bones[i].coordinate_matrix[2][0],
-                     md->bones[i].coordinate_matrix[2][1],
-                     md->bones[i].coordinate_matrix[2][2],
-                     parent, md->bones[i].num_children);
-  }
-  fprintf(file, "\n");
-
-  // Verticies
-  for (size_t i = 0; i < md->num_vertices; i++) {
-    fprintf(file, "v %f %f %f %d:%f %d:%f %d:%f %d:%f\n",
-            md->vertices[i].vertex[X],
-            md->vertices[i].vertex[Y],
-            md->vertices[i].vertex[Z],
-            md->vertices[i].bone_ids[X],
-            md->vertices[i].weights[X],
-            md->vertices[i].bone_ids[Y],
-            md->vertices[i].weights[Y],
-            md->vertices[i].bone_ids[Z],
-            md->vertices[i].weights[Z],
-            md->vertices[i].bone_ids[W],
-            md->vertices[i].weights[W]);
-  }
-  for (size_t i = 0; i < md->num_vertices; i++) {
-    fprintf(file, "vt %f %f\n",
-            md->vertices[i].tex_coord[X],
-            md->vertices[i].tex_coord[Y]);
-  }
-  for (size_t i = 0; i < md->num_vertices; i++) {
-    fprintf(file, "vn %f %f %f\n",
-            md->vertices[i].normal[X],
-            md->vertices[i].normal[Y],
-            md->vertices[i].normal[Z]);
-  }
-
-  // Face indicies
-  for (size_t i = 0; i < md->num_indices / 3; i++) {
-    fprintf(file, "f %d/%d/%d %d/%d/%d %d/%d/%d\n",
-            md->indices[(i*3)] + 1,
-            md->indices[(i*3)] + 1,
-            md->indices[(i*3)] + 1,
-            md->indices[(i*3)+1] + 1,
-            md->indices[(i*3)+1] + 1,
-            md->indices[(i*3)+1] + 1,
-            md->indices[(i*3)+2] + 1,
-            md->indices[(i*3)+2] + 1,
-            md->indices[(i*3)+2] + 1);
-  }
-  fprintf(file, "\n");
-
-  // Colliders
-  for (size_t i = 0; i < md->num_colliders; i++) {
-    if (md->colliders[i].type == POLY) {
-      vec3 verts[8];
-      int root_bone = md->collider_bone_links[i];
-      // Collider verts in .obj files are given in entity space, but are stored
-      // in bone face in engine, so we must convert them back to entity space
-      if (root_bone != -1) {
-        mat4 bone_to_entity = GLM_MAT4_IDENTITY_INIT;
-        glm_mat4_ins3(md->bones[root_bone].coordinate_matrix, bone_to_entity);
-        glm_vec4(md->colliders[i].data.center_of_mass, 1.0, bone_to_entity[3]);
-
-        // Convert collider verticies to bone space
-        for (int j = 0; j < 8; j++) {
-          glm_mat4_mulv3(bone_to_entity, md->colliders[i].data.verts[j], 1.0,
-                         verts[j]);
-        }
-      }
-
-      fprintf(file, "hp %d %d %d %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f %f\n",
-              md->colliders[i].category,
-              md->collider_bone_links[i]-1,
-              md->colliders[i].data.num_used,
-              verts[0][X],
-              verts[0][Y],
-              verts[0][Z],
-              verts[1][X],
-              verts[1][Y],
-              verts[1][Z],
-              verts[2][X],
-              verts[2][Y],
-              verts[2][Z],
-              verts[3][X],
-              verts[3][Y],
-              verts[3][Z],
-              verts[4][X],
-              verts[4][Y],
-              verts[4][Z],
-              verts[5][X],
-              verts[5][Y],
-              verts[5][Z],
-              verts[6][X],
-              verts[6][Y],
-              verts[6][Z],
-              verts[7][X],
-              verts[7][Y],
-              verts[7][Z]);
-    } else {
-      fprintf(file, "hs %d %d %f %f %f %f\n",
-              md->colliders[i].category,
-              md->collider_bone_links[i]-1,
-              md->colliders[i].data.center[X],
-              md->colliders[i].data.center[Y],
-              md->colliders[i].data.center[Z],
-              md->colliders[i].data.radius);
-    }
-    for (int j = 0; j < md->colliders[i].num_dofs; j++) {
-      int dof_type = md->colliders[i].dofs[j][W];
-#ifdef __linux__
-      fprintf(file, "dof %ld %d %f %f %f\n",
-#else
-      fprintf(file, "dof %lld %d %f %f %f\n",
-#endif
-              i, dof_type,
-              md->colliders[i].dofs[j][X],
-              md->colliders[i].dofs[j][Y],
-              md->colliders[i].dofs[j][Z]);
-    }
-  }
-  fprintf(file, "\n");
-
-  // Animations
-  for (size_t i = 0; i < md->num_animations; i++) {
-#ifdef __linux__
-    fprintf(file, "a %ld\n", md->animations[i].duration);
-#else
-    fprintf(file, "a %lld\n", md->animations[i].duration);
-#endif
-    K_CHAIN *chains = md->animations[i].keyframe_chains;
-    for (size_t j = 0; j < md->animations[i].num_chains; j++) {
-      if (chains[j].type == LOCATION) {
-        fprintf(file, "cl %d\n", chains[j].b_id);
-      } else if (chains[j].type == ROTATION) {
-        fprintf(file, "cr %d\n", chains[j].b_id);
-      } else {
-        fprintf(file, "cs %d\n", chains[j].b_id);
-      }
-      for (size_t k = 0; k < chains[j].num_frames; k++) {
-        if (chains[j].type == ROTATION) {
-          fprintf(file, "kp %d %f %f %f %f\n",
-                  chains[j].chain[k].frame,
-                  chains[j].chain[k].offset[X],
-                  chains[j].chain[k].offset[Y],
-                  chains[j].chain[k].offset[Z],
-                  chains[j].chain[k].offset[W]);
-        } else {
-          fprintf(file, "kp %d %f %f %f\n",
-                  chains[j].chain[k].frame,
-                  chains[j].chain[k].offset[X],
-                  chains[j].chain[k].offset[Y],
-                  chains[j].chain[k].offset[Z]);
-        }
-      }
-    }
-    fprintf(file, "\n");
-  }
-
-  fclose(file);
-  return 0;
-}
-
-int write_model_bin(MODEL_DATA *md, char *path) {
-  FILE *file = fopen(path, "wb");
-  if (!file) {
-    return -1;
-  }
-  size_t total_chains = 0;
-  size_t total_keyframes = 0;
-  size_t total_frames = 0;
-  // Compute totals for animation
-  for (size_t i = 0; i < md->num_animations; i++) {
-    total_chains += md->animations[i].num_chains;
-    total_frames += (md->animations[i].duration *
-                     md->animations[i].num_chains);
-    for (int j = 0; j < md->animations[i].num_chains; j++) {
-      total_keyframes += md->animations[i].keyframe_chains[j].num_frames;
-    }
-  }
-
-  fwrite(&md->num_bones, sizeof(size_t), 1, file);
-  fwrite(&md->num_colliders, sizeof(size_t), 1, file);
-  fwrite(&md->num_vertices, sizeof(size_t), 1, file);
-  fwrite(&md->num_indices, sizeof(size_t), 1, file);
-  fwrite(&md->num_animations, sizeof(size_t), 1, file);
-
-  fwrite(&total_chains, sizeof(size_t), 1, file);
-  fwrite(&total_keyframes, sizeof(size_t), 1, file);
-  fwrite(&total_frames, sizeof(size_t), 1, file);
-
-  // TODO Implement material saving
-  int material_flag = 0;
-  fwrite(&material_flag, sizeof(material_flag), 1, file);
-
-  fwrite(md->bones, sizeof(BONE), md->num_bones, file);
-  fwrite(md->bone_collider_links, sizeof(int), md->num_bones, file);
-
-  fwrite(md->colliders, sizeof(COLLIDER), md->num_colliders, file);
-  fwrite(md->collider_bone_links, sizeof(int), md->num_colliders, file);
-
-  fwrite(md->vertices, sizeof(VBO), md->num_vertices, file);
-  fwrite(md->indices, sizeof(int) * 3, md->num_indices, file);
-
-  for (size_t i = 0; i < md->num_animations; i++) {
-    fwrite(&(md->animations[i].num_chains), sizeof(size_t), 1, file);
-    fwrite(&(md->animations[i].duration), sizeof(size_t), 1, file);
-    for (size_t j = 0; j < md->animations[i].num_chains; j++) {
-      K_CHAIN cur = md->animations[i].keyframe_chains[j];
-      fwrite(&(cur.b_id), sizeof(unsigned int), 1, file);
-      fwrite(&(cur.type), sizeof(C_TYPE), 1, file);
-      fwrite(&(cur.num_frames), sizeof(size_t), 1, file);
-      for (size_t k = 0; k < cur.num_frames; k++) {
-        fwrite(cur.chain[k].offset, sizeof(float), 4, file);
-        fwrite(&(cur.chain[k].frame), sizeof(int), 1, file);
-      }
-    }
-  }
-
-  fclose(file);
-  return 0;
-}
